+++ app/Modules/LeadManagement/Routes/api.php
@@ -4,9 +4,9 @@ use App\Modules\LeadManagement\Controllers\Api\LeadController;
 use Illuminate\Support\Facades\Route;
 
 Route::middleware(['api', 'auth:sanctum'])->prefix('api')->group(function () {
-    Route::get('leads', [LeadController::class, 'index']);
-    Route::post('leads', [LeadController::class, 'store']);
-    Route::get('leads/{lead}', [LeadController::class, 'show']);
-    Route::put('leads/{lead}', [LeadController::class, 'update']);
-    Route::delete('leads/{lead}', [LeadController::class, 'destroy']);
+    Route::get('leads', [LeadController::class, 'index'])->middleware('permission:leads.view');
+    Route::post('leads', [LeadController::class, 'store'])->middleware('permission:leads.create');
+    Route::get('leads/{lead}', [LeadController::class, 'show'])->middleware('permission:leads.view');
+    Route::put('leads/{lead}', [LeadController::class, 'update'])->middleware('permission:leads.update');
+    Route::delete('leads/{lead}', [LeadController::class, 'destroy'])->middleware('permission:leads.delete');
 });
