+++ app/Modules/Users/Routes/api.php
@@ -4,8 +4,10 @@ use App\Modules\Users\Controllers\Api\UserController;
 use Illuminate\Support\Facades\Route;
 
 Route::middleware(['api', 'auth:sanctum'])->prefix('api')->group(function () {
-    Route::get('users', [UserController::class, 'index']);
-    Route::get('users/{user}', [UserController::class, 'show']);
-    Route::put('users/{user}', [UserController::class, 'update']);
-    Route::delete('users/{user}', [UserController::class, 'destroy']);
+    Route::get('users/meta', [UserController::class, 'meta']);
+    Route::get('users', [UserController::class, 'index'])->middleware('permission:users.view');
+    Route::post('users', [UserController::class, 'store'])->middleware('permission:users.create');
+    Route::get('users/{user}', [UserController::class, 'show'])->middleware('permission:users.view');
+    Route::put('users/{user}', [UserController::class, 'update'])->middleware('permission:users.update');
+    Route::delete('users/{user}', [UserController::class, 'destroy'])->middleware('permission:users.delete');
 });
